# swaggerdocs.json
{
	"openapi": "3.0.1",
	"info": {
		"title": "OpenAPI definition",
		"version": "v0"
	},
	"servers": [
		{
			"url": "http://localhost:8081",
			"description": "Generated server url"
		}
	],
	"paths": {
		"/addbook": {
			"get": {
				"tags": [
					"book-controller"
				],
				"operationId": "getAllBooks",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"*/*": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Book"
									}
								}
							}
						}
					}
				}
			},
			"put": {
				"tags": [
					"book-controller"
				],
				"operationId": "updateBook",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Book"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"*/*": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				}
			},
			"post": {
				"tags": [
					"book-controller"
				],
				"operationId": "addBook",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Book"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"*/*": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				}
			}
		},
		"/addbook/{bid}": {
			"delete": {
				"tags": [
					"book-controller"
				],
				"operationId": "deleteBook",
				"parameters": [
					{
						"name": "bid",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int32"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"*/*": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				}
			}
		}
	},
	"components": {
		"schemas": {
			"Book": {
				"type": "object",
				"properties": {
					"bid": {
						"type": "integer",
						"format": "int32"
					},
					"name": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					}
				}
			}
		}
	}
}